pyproject.toml
[tool.poetry]
name = "handwriting"
version = "0.1.0"
description = ""
authors = ["Marc Mezger <marc.mezger@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.17.0"
aleph-alpha-client = "^2.7.1"
python-dotenv = "^0.21.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
